assets: use AnySceneImporter for loading general mesh data.
This makes it support not only glTF, but also everything Assimp
knows about and also basically anything else AnySceneImporter can
detect, if a corresponding plugin is found.

Some mesh formats supply just mesh data and no scene hierarchy, in that
case take a wild guess and import the first mesh and slap the first
material on it.